Refactor/ viewPager에 들어가는 Fragment들의 Model 통일
