+++ .idea/misc.xml
@@ -13,6 +13,7 @@
         <entry key="app/src/main/res/layout/fragment_community.xml" value="0.28072916666666664" />
         <entry key="app/src/main/res/layout/fragment_community_detail.xml" value="0.28072916666666664" />
         <entry key="app/src/main/res/layout/fragment_curriculum.xml" value="0.2796875" />
+        <entry key="app/src/main/res/layout/fragment_detail.xml" value="0.28072916666666664" />
         <entry key="app/src/main/res/layout/fragment_login.xml" value="0.2764102564102564" />
         <entry key="app/src/main/res/layout/fragment_login_main.xml" value="0.2764102564102564" />
         <entry key="app/src/main/res/layout/fragment_member.xml" value="0.2796875" />
