+++ app/src/main/java/com/headingWarm/usha/community/detailpages/viewpagers/MemberFragment.kt
@@ -36,13 +36,14 @@ class MemberFragment: Fragment() {
         binding = FragmentMemberBinding.inflate(inflater).apply {
             lifecycleOwner = this@MemberFragment
             viewModel = ViewModelProvider(this@MemberFragment,
-                FragmentMemberViewModel.FragmentMemberVMFac(FragmentMemberModel(community!!, resources.displayMetrics.widthPixels, findNavController()))
-                ).get(FragmentMemberViewModel::class.java)
+                FragmentMemberViewModel.FragmentMemberVMFac(DetailModel(community!!, resources.displayMetrics.widthPixels, findNavController())
+                    .apply { setBitmap(community.mentor_img) }
+                )).get(FragmentMemberViewModel::class.java)
         }
         return binding.root
     }
 
-    class FragmentMemberViewModel(val model: FragmentMemberModel): ViewModel(){
+    class FragmentMemberViewModel(val model: DetailModel): ViewModel(){
 
         val image: LiveData<Bitmap> get() = model.image
 
@@ -50,48 +51,12 @@ class MemberFragment: Fragment() {
             model.clickFloatingBtn()
         }
 
-        class FragmentMemberVMFac(val model: FragmentMemberModel): ViewModelProvider.Factory{
+        class FragmentMemberVMFac(val model: DetailModel): ViewModelProvider.Factory{
             override fun <T : ViewModel> create(modelClass: Class<T>): T {
                 return FragmentMemberViewModel(model) as T
             }
         }
 
-    }
-
-    class FragmentMemberModel(val community: Community, val displayWidth: Int, val navController: NavController){
-        val image = MutableLiveData<Bitmap>()
-
-        init {
-            setBitmap(community.mentor_img)
-        }
-
-        fun setBitmap(url: String){
-            Observable.just(url)
-                .subscribeOn(Schedulers.io())
-                .observeOn(Schedulers.io())
-                .map {
-                    var iStream = URL(url).openStream()
-                    var bmp = BitmapFactory.decodeStream(iStream)
-                    var height = (displayWidth.toFloat()/bmp.width.toFloat())*bmp.height
-                    Bitmap.createScaledBitmap(bmp,displayWidth, height.toInt(),false)
-                }
-                .observeOn(AndroidSchedulers.mainThread())
-                .subscribe {image.value = it}
-        }
-
-        fun clickFloatingBtn(){
-            if(MyApplication.loginInfo.loginned){
-                // 로그인 되어 있으면 가입 페이지로 보낸다.
-                val bundle = Bundle().apply {
-                    putSerializable("community",community)
-                }
-                navController.navigate(R.id.reservation, bundle)
-            }
-            else{
-                // 로그인 안되어 있으면 로그인 페이지로 보낸다.
-                navController.navigate(R.id.loginMain)
-            }
-        }
     }
 
 }
